requirements.txt
discord.py[voice]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
coloredlogs
